+++ src/components/navigation-bar.tsx
@@ -22,8 +22,8 @@ export function NavigationBar() {
 
   return (
     <div
-      className={`z-100 fixed right-6 bottom-12 md:bottom-18 lg:bottom-1/2 lg:translate-y-1/2 flex flex-col items-center text-white max-w-min rounded-lg ${
-        isExpanded ? "opacity-100" : "opacity-70"
+      className={`z-100 fixed right-6 bottom-12 md:bottom-18 lg:bottom-1/2 lg:translate-y-1/2 text-white max-w-min rounded-lg ${
+        isExpanded ? "opacity-100" : "opacity-40"
       } transition-opacity duration-300`}
       onMouseEnter={() => setIsExpanded(true)}
       onMouseLeave={() => setIsExpanded(false)}
@@ -41,7 +41,7 @@ export function NavigationBar() {
           ))}
         </div>
       ) : (
-        <div className="flex flex-col items-center p-4 space-y-1 bg-background rounded-lg">
+        <div className="flex flex-row items-center p-4 space-x-1 lg:space-x-0 lg:space-y-1 lg:flex-col bg-background rounded-lg">
           {navItems.map((item) => (
             <div
               key={item.key} // Changed from index to item.key