+++ src/components/ui/card.tsx
@@ -7,7 +7,7 @@ function Card({ className, ...props }: React.ComponentProps<"div">) {
     <div
       data-slot="card"
       className={cn(
-        "bg-card text-card-foreground flex flex-col gap-2 md:gap-6 rounded-xl border py-4 md:py-6 shadow-sm overflow-y-auto",
+        "bg-card text-card-foreground flex flex-col gap-2 2xl:gap-6 rounded-xl border py-4 2xl:py-6 shadow-sm overflow-y-auto",
         className
       )}
       {...props}
@@ -20,7 +20,7 @@ function CardHeader({ className, ...props }: React.ComponentProps<"div">) {
     <div
       data-slot="card-header"
       className={cn(
-        "@container/card-header grid auto-rows-min grid-rows-[auto_auto] items-start gap-1 md:gap-1.5 px-4 md:px-6 has-data-[slot=card-action]:grid-cols-[1fr_auto] [.border-b]:pb-2 [.border-b]:md:pb-6",
+        "@container/card-header grid auto-rows-min grid-rows-[auto_auto] items-start gap-1 2xl:gap-1.5 px-4 2xl:px-6 has-data-[slot=card-action]:grid-cols-[1fr_auto] [.border-b]:pb-2 [.border-b]:2xl:pb-6",
         className
       )}
       {...props}
@@ -65,7 +65,7 @@ function CardContent({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="card-content"
-      className={cn("px-4 md:px-6", className)}
+      className={cn("px-4 2xl:px-6", className)}
       {...props}
     />
   );
@@ -76,7 +76,7 @@ function CardFooter({ className, ...props }: React.ComponentProps<"div">) {
     <div
       data-slot="card-footer"
       className={cn(
-        "flex items-center px-2 md:px-6 [.border-t]:pt-6",
+        "flex items-center px-2 2xl:px-6 [.border-t]:pt-6",
         className
       )}
       {...props}