+++ change_log.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## 2025-05-01
 
-- Changed footer and header to keep consistent layout and style
+- Changed interface to keep consistent layout and style
+- Added warning on too little height to render content properly
+- Saved space for mobile rotated screen
 
 ## 2025-04-29